Pasaje de double a BigDecimal para representar dinero

+++ src/logica/valueobjects/VOBoletoEspecialEntrada.java
@@ -1,15 +1,18 @@
 package logica.valueobjects;
 
+import java.math.BigDecimal;
+
 public class VOBoletoEspecialEntrada extends VOBoletoEntrada {
 
-    private double descuento;
+    private BigDecimal descuento;
 
-    public VOBoletoEspecialEntrada(double descuento,String codigoExcursion,int edad,String procedencia,String celular) {
+    public VOBoletoEspecialEntrada(BigDecimal descuento, String codigoExcursion,
+    		int edad, String procedencia, String celular) {
         super( codigoExcursion,edad,procedencia,celular);
         this.descuento = descuento;
     }
 
-    public double getDescuento() {
+    public BigDecimal getDescuento() {
         return descuento;
     }
